Finalizando a Criacao de API, README, requirements.txt e outras melhorias e mudancas que precisavam ser feitas

+++ main.py
@@ -1,32 +1,33 @@
-from fastapi import FastAPI, HTTPException
+from fastapi import FastAPI, HTTPException, UploadFile, File
 from fastapi.responses import JSONResponse
+import ia_preditc
 
 import schemas
 from database import crud
 
-
 app = FastAPI()
 
 
-@app.post("/users", status_code=201, response_model=schemas.User)
+@app.post(
+    "/users",
+    status_code=201,
+    response_model=schemas.User,
+    summary="Criar um novo usuario",
+)
 def create_user(user: schemas.UserCreate):
-    db_user = crud.create_user(user)
-    return db_user
+    """
+    Area para criar um novo usuario \n
+    ! Se atente em colocar as entradas e informacoes corretas !
+    """
+    db_users = crud.create_user(user)
+    return db_users
 
 
-@app.get("/users/{user_id}", response_model=list[schemas.User])
-def get_user(user_id: int):
+@app.get("/users/{user_id}", response_model=schemas.User)
+def get_user_by_id(user_id: int):
     user = crud.read_user(user_id)
     if not user:
-        raise HTTPException(status_code=404, detail="Store not found")
-    return user
-
-
-@app.get("/users/{cpf}", response_model=schemas.User)
-def get_user_id(cpf: str):
-    user = crud.read_user_by_cpf(cpf)
-    if not user:
-        raise HTTPException(status_code=404, detail="CPF not found")
+        raise HTTPException(status_code=404, detail="user not found")
     return user
 
 
@@ -34,43 +35,93 @@ def get_user_id(cpf: str):
 def update_user(user_id: int, user: schemas.UserCreate):
     db_user = crud.update_user(user_id, user)
     if not db_user:
-        raise HTTPException(status_code=404, detail="User not found")
+        raise HTTPException(status_code=404, detail="user not found")
+
     return db_user
 
 
-@app.delete("/users/{user_id}")
-def delete_user(user_id: int):
-    deleted_user = crud.delete_user(user_id)
-    if not deleted_user:
-        raise HTTPException(status_code=404, detail="User not found")
-    return {"detail": "User successfully removed"}
+@app.delete("/users/{user_id}", response_model=schemas.User)
+def delete_user_by_id(user_id: int):
+    user = crud.read_user(user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="user not found")
+
+    crud.delete_user(user_id)
+
+    message = {"message": "User deletado com sucesso"}
+    return JSONResponse(content=message)
 
 
 @app.post("/users/{user_id}/bikes/", status_code=201, response_model=schemas.Bike)
-def add_bike(user_id: int, bike: schemas.BikeCreate):
-    new_bike = crud.create_bike(user_id, bike)
-    return new_bike
+def create_bike(user_id: int, bike: schemas.BikeCreate):
+    user = crud.read_user(user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="user not found")
+
+    db_bike = crud.create_bike(user_id, bike)
+    return db_bike
 
 
 @app.get("/bikes/{bike_id}", response_model=schemas.Bike)
 def get_bike(bike_id: int):
     bike = crud.read_bike(bike_id)
     if not bike:
-        raise HTTPException(status_code=404, detail="Bike not found")
+        raise HTTPException(status_code=404, detail="bike not found")
     return bike
 
 
 @app.put("/bikes/{bike_id}", response_model=schemas.Bike)
-def update_bike(bike_id: int, bike: schemas.BikeCreate):
-    updated_bike = crud.update_bike(bike_id, bike)
-    if not updated_bike:
-        raise HTTPException(status_code=404, detail="Bike not found")
-    return updated_bike
+def update_bike(bike_id: int, user_id: int, bike: schemas.BikeCreate):
+    db_bike = crud.update_bike_in_db(user_id, bike_id, bike)
+    if not db_bike:
+        raise HTTPException(status_code=404, detail="Bicicleta não encontrada")
+
+    return db_bike
 
 
 @app.delete("/bikes/{bike_id}", response_model=schemas.Bike)
-def remove_bike(bike_id: int):
-    deleted_bike = crud.delete_bike(bike_id)
-    if not deleted_bike:
-        raise HTTPException(status_code=404, detail="Bike not found")
-    return deleted_bike
+def delete_bike_by_id(bike_id: int):
+    bike = crud.read_bike(bike_id)
+    if not bike:
+        raise HTTPException(status_code=404, detail="bike not found")
+
+    crud.delete_bike(bike_id)
+
+    message = {"message": "bike deleted"}
+    return JSONResponse(content=message)
+
+
+@app.get("/users/{user_id}/bikes/", response_model=list[schemas.Bike])
+def get_bike_from_user(user_id: int):
+    user = crud.read_user(user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="user not found")
+    bikes = crud.read_all_bikes_from_user(user_id)
+    return bikes
+
+
+@app.get("/bikes", response_model=list[schemas.Bike])
+def get_bike():
+    bikes = crud.read_all_bikes()
+    return bikes
+
+
+@app.post("/predict-bike-image", response_model=schemas.ImagePrediction)
+async def predict_bike_image(
+    file: UploadFile = File(...),
+):
+    image_path, prediction_image_path, predicted_class, confidence = ia_preditc.cv_api(
+        file.file
+    )
+
+    # Salvar no banco de dados
+    db_prediction = crud.create_image_prediction(
+        schemas.ImagePredictionCreate(
+            image_path=image_path,
+            prediction_image_path=prediction_image_path,
+            predicted_class=predicted_class,
+            confidence=confidence,
+        ),
+    )
+
+    return db_prediction
